mybiseoServer/requirements.txt
bleach==1.5.0
boto==2.48.0
bz2file==0.98
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.4
gensim==2.3.0
gunicorn==19.7.1
html5lib==0.9999999
idna==2.6
JPype1-py3==0.5.5.2
konlpy==0.4.4
Markdown==2.6.9
numpy==1.13.1
protobuf==3.4.0
pytz==2017.2
requests==2.18.4
scipy==0.19.1
six==1.10.0
smart-open==1.5.3
tensorflow==1.3.0
tensorflow-tensorboard==0.1.5
urllib3==1.22
virtualenv==15.1.0
Werkzeug==0.12.2
